+++ nodes/Inistate/Inistate.node.ts
@@ -33,8 +33,8 @@ export class Inistate implements INodeType {
 		credentials: [{ name: 'inistateApi', required: true }],
 		properties: [
 			{
-				displayName: 'The API environment is selected in the Inistate credential.',
-				name: 'environmentNotice',
+				displayName: 'The API host is configured in the Inistate credential.',
+				name: 'hostNotice',
 				type: 'notice',
 				default: '',
 			},